making agent thread persistent
In the process of making the agent thread itself persistent for
persistent agents. It turns out that in windows, each thread holds its
fair share of settings and resources, and if a thread with modified
resources terminates, it can affect the rest of the program. As such, we
will modify all agents labeled 'persistent' such that not only is the
data persistent, but the original agent thread is persistent as well
until the agent is terminated by a third party.