Added some initial code for the agent_data_share
The agent_data_share will be a data structure carried by the agents
which allows them to expose chunks of data. This was created out of
necessity; the existing functions such as "MC_AgentVariableSave" only
saves the variables upon agent termination, but in some cases, we need
to access agent data before the agent terminates. These functions will
allow agents to "post" certain named pieces of information allowing the
agency or other agents to access those pieces of data while the agent is
still running.